Refactoring of StDebugger to instanciate a debugger action model when we give the debugger a debug session.
This is necessary to avoid the action model's lazy initialization in the debugger by providing it the debugger's session, as this session will eventually be accessed via the action model.
For now, `StDebugger >> session:` sets the session for the session holder AND the action model so that I can do the refactoring step by step but the session holder will eventually be removed.